fix: give acquisition timeout in seconds.

+++ silq/instrument_interfaces/keysight/M3300A_DIG_interface.py
@@ -236,7 +236,7 @@ class M3300A_DIG_Interface(InstrumentInterface):
 
             # Set an acquisition timeout to be 10% after the last pulse finishes.
             # NOTE: time is defined in seconds
-            controller.read_timeout(int(t_final * 1.1))
+            controller.read_timeout(t_final * 1.1)
 
     def start(self):
         self._acquisition_controller.pre_start_capture()
